+++ CHANGELOG.md
@@ -19,6 +19,11 @@
 * Fixed a bug in `--experimental-reset-state-function` support for heap reset.
   [#4665](https://github.com/wasm-bindgen/wasm-bindgen/pull/4665)
 
+### Changed
+
+* Unconditionally use `TextEncoder`/`TextDecoder` for string encoding/decoding.
+  [#TODO]
+
 --------------------------------------------------------------------------------
 
 ## [0.2.103](https://github.com/wasm-bindgen/wasm-bindgen/compare/0.2.102...0.2.103)